+++ source/app/service-providers/config-provider.js
@@ -100,7 +100,7 @@ module.exports = class ConfigProvider extends EventEmitter {
       // Visible attachment filetypes
       'attachmentExtensions': COMMON_DATA.attachmentExtensions,
       // UI related options
-      'darkTheme': false, // TODO DEPRECATED to be renamed to darkMode
+      'darkMode': false,
       'alwaysReloadFiles': false, // Should Zettlr automatically load remote changes?
       'autoDarkMode': 'off', // Possible values: 'off', 'system', 'schedule', 'auto'
       'autoDarkModeStart': '22:00', // Switch into dark mode at this time